Switched to python to download gatb-core

+++ core/kover/dataset/tools/kmer_tools/install.sh
@@ -1,26 +1,42 @@
 # Remove any prior build
 rm -rf build
 
-#Removing gatb-core previous version
+# Preparing gatb-core folder
 cd thirdparty
-rm -rf gatb-core
+if [ -d "gatb-core" ]; then rm -rf gatb-core; fi
 mkdir gatb-core
 cd gatb-core
 
-#Downloading gatb-core
-VERSION=v1.2.2
-URL=https://github.com/GATB/gatb-core/archive/$VERSION.tar.gz
-wget $URL
-tar -zxf $VERSION.tar.gz --strip-components=1
-rm $VERSION.tar.gz
+#Download gatb-core
+echo 'Downloading gatb-core.'
+case "$(python --version 2>&1)" in
+    *" 2."*)
+        python -c 'import urllib; urllib.urlretrieve ("https://github.com/GATB/gatb-core/archive/v1.2.2.tar.gz", "gatbcore.tar.gz")'
+        ;;
+	*" 3."*)
+		python -c 'import urllib.request; urllib.request.urlretrieve ("https://github.com/GATB/gatb-core/archive/v1.2.2.tar.gz", "gatbcore.tar.gz")'
+	;;
+    *)
+        echo "Error: Python 2 or 3 is not installed"
+        ;;
+esac
+tar -zxf gatbcore.tar.gz --strip-components=1
+rm gatbcore.tar.gz
 cd ../..
-
 # Create build directory
 mkdir  build
 cd build
 
 # Prepare the makefile
+echo 'Compiling gatb-core.'
 cmake .. -Wno-dev
 
 # Run the makefile to build dsk, multidsk and dsk2kover tools:
 make -j8
+
+# Clean up gatb-core source files
+echo 'Cleaning up gatb-core source files.'
+cd ..
+cd thirdparty
+rm -rf gatb-core
+cd ..
